requirements.txt
transformers==4.41.0
datasets==2.16.0
torch==2.2.0
accelerate==0.25.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
chromadb==0.4.3
openai==1.9.2
streamlit==1.26.0
gradio==3.43.0
numpy==1.25.2
pandas==2.1.1
PyPDF2==3.0.1
python-dotenv==1.0.1
tqdm==4.66.1
scikit-learn==1.3.1
nltk==3.8.1
PyMuPDF == 1.26.4
langchain
sentence-transformers
peft
bitsandbytes==0.39.1
